[VarDumper] replace VarDumperTestCase by trait
