+++ src/Symfony/Component/VarDumper/Tests/Test/VarDumpTestTraitTest.php
@@ -9,8 +9,32 @@
  * file that was distributed with this source code.
  */
 
-// Skipping trait tests for PHP < 5.4
-if (version_compare(PHP_VERSION, '5.4.0-dev', '>=')) {
-    require 'VarDumpTestTraitRequire54.php';
-}
+namespace Symfony\Component\VarDumper\Tests\Test;
 
+use Symfony\Component\VarDumper\Test\VarDumperTestTrait;
+
+class VarDumperTestTraitTest extends \PHPUnit_Framework_TestCase
+{
+    use VarDumperTestTrait;
+
+    public function testItComparesLargeData()
+    {
+        $howMany = 700;
+        $data = array_fill_keys(range(0, $howMany), array('a', 'b', 'c', 'd'));
+
+        $expected = sprintf("array:%d [\n", $howMany + 1);
+        for ($i = 0; $i <= $howMany; ++$i) {
+            $expected .= <<<EODUMP
+  $i => array:4 [
+    0 => "a"
+    1 => "b"
+    2 => "c"
+    3 => "d"
+  ]\n
+EODUMP;
+        }
+        $expected .= "]\n";
+
+        $this->assertDumpEquals($expected, $data);
+    }
+}
